Project/requirements.txt
kears==2.4.3
tensorflow==2.3.0
glob2==0.7
numpy==1.18.5
pickleshare==0.7.5
pandas==1.3.3
h5py==2.10.0
matplotlib==3.4.3
scipy==1.4.1
opencv-python==4.1.2.30
opencv-contrib-python==4.1.2.30
seaborn==0.11.1
scikit-learn==0.23.1
joblib==1.0.1
Flask==2.0.3
Jinja2==0.17.2
Markdown==3.2.2
Werkzeug==2.0.3
tqdm==4.64.0
db-sqlite3==0.0.1
efficientnet==1.1.1
efficientnet-pytorch==0.6.3
keras-efficientnets==0.1.7